Using Selenium (Headless Chrome) instead of Capybara Webkit
Capybara Wekit relies on Qt which at this point is equivalent of a 6 to 7 year old Safari browser. Doesn't even support flexbox.

Switching over to Selenium with Headless Chrome. Original preference was Selenium with Firefox but that requires specific versions of browser and driver to be installed locally and on the CI.