Renaming `Photos::SourceFileController` -> `RawPhotosController`
The namespacing was previously causing an error in autoloading because it expected a nested module instead of a namespaced constant.
However a nested module requires a blank file at the `app/controllers/` root to define the parent namespace, which feels messy.

The above autoloading was causing the following issues:

* The first photo rendered would often not load because the first call would raise an autoload error
* The feature specs ran into the same issue as above and would output a large stack trace

Ultimately the new name is a bit clearer and all around simpler.