Mock `Date.now()` more than once
The test was failing because `Date.now()` was only being mocked once and
the actual `setTimeout` on component load was using real time.

It could be due to a recent change in react or jest or any library. To
be safe we just mock `Date.now()` infinitely instead of just once.