Skip mounting node if `mountNode` is not found
This occurs on pages where the component tries to mount itself but we
don't want to load the component on that page.

e.g. `ActionsNotifications` tried to mount itself on the
`Collections#show` page but that page has no DOM element to mount on
because we never call `react_component()`

This was causing an error on `mountNode.getAttribute` because
`mountNode` was `null`