[Girder] Using the `i18n-js` with webpack
The `i18n-js` library has three components:

* A frontend JS library tha provides an `I18n` object
* A frontend translations file that loads and sets all the required translations
* A backend middleware that autmatically regenerates the frontend translation file if needed

This commit exports the frontend JS library as it's own "pack" so that it can be set globally before loading any other JS code. We disable the default option to export the I18n JS file.

We then load the translations file. It is configured to automatically generate only on `development` and `test`, and we configure a custom directory (`public/i18n/translations.js`) to output it to.