[Girder] Creating shared `admin/_filter_warning` partial
Previously the `admin/audits#index` page used a filter warning when the
result set was being filtered.

Moving that logic to a shared partial so other admin pages can use it.

The downside is that they all share a translation so it's a little bit
more inflexible on wording, but it should work for the majority of basic
cases.