[Girder] Allow passing a function to `<Modal>` to close the modal
Up till now we have closed the modal by unmounting whatever component was on the `#modal` DOM node.

This isn't great practice. Allow the component to unmount by passing in a function from a parent that controls its rendering state.

The old approach is still used as a fallback.

Also updated `<ModalForm>` and `<ModalPrompt>` to pass through the
function to `<Modal>`