[Girder] Send user to main login page if native auth is disabled
By default when we send an invitation mailer we direct the user to the
registration url. This URL is blocked when native auth is disabled, and
the user gets redirect with a flash message.

Instead, if native auth is disabled just send the user straight to the
login page which will have the OAuth login button.