[Girder] Adding "modified" column to Audit Log
The audit log at `/admin/audits` lists all activity across the app.

We already display the "modifier" - the person that takes the action -
for each audit. This is easy since it's built into the `Audit` record as
a relation.

This change adds a new "modified" column that lists the user that was
modified as a result of that change. Sometimes there is no user modified
(e.g. a `Skill` is edited).

Determining the "modified" `User` is non-trivial because it requires
looking at the individual `Audit` record and determining the `User`
impacted.

e.g.

* For "comment created" Audits, the modified user is the `Comment#owner`
* For "role was added" Audits, the modified user is the `Role#user_id`

Similar to the existing "modifier" column, we also add in the ability to
click on a User's name and have it filter on all records for that user.
Again, this is non-trivial for the "modified" column because we have to
scrape together all audits that impact a particular `User` using a
concatenated SQL query.

Lastly, we make some minor updates to the copy that appears when the
viewer is viewing a filtered list.