[Girder] [BUG] Notify inviter only once
The `UserInvitationMailer#notify_inviter_of_completion` is a
notification to an inviter that the person they invited has signed up
for the application.

Ideally this should only be sent once - the first time someone signs up.
In the case of OAuth, "sign in" and "sign up" are equivalent and we
don't distinguish them, so we end up sending this notification to an
inviter every time someone they invited in the past signs in.

This change starts tracking whether a user was created ("sign up") or
not ("sign in"), and we use that information to only send the above
maile once, on sign up.